+++ src/Adapter/Supplier/Repository/SupplierRepository.php
@@ -26,7 +26,7 @@
 
 declare(strict_types=1);
 
-namespace PrestaShop\PrestaShop\Adapter\Supplier;
+namespace PrestaShop\PrestaShop\Adapter\Supplier\Repository;
 
 use PrestaShop\PrestaShop\Adapter\AbstractObjectModelRepository;
 use PrestaShop\PrestaShop\Core\Domain\Supplier\Exception\SupplierNotFoundException;
