+++ src/PrestaShopBundle/Resources/config/services/adapter/product.yml
@@ -316,7 +316,7 @@ services:
         class: PrestaShop\PrestaShop\Adapter\Product\ProductSupplierValidator
         arguments:
             - '@prestashop.adapter.product.product_repository'
-            - '@prestashop.adapter.supplier.supplier_repository'
+            - '@prestashop.adapter.supplier.repository.supplier_repository'
             - '@prestashop.adapter.currency.repository.currency_repository'
 
     prestashop.adapter.product.repository.product_supplier_repository:
@@ -328,5 +328,5 @@ services:
         class: PrestaShop\PrestaShop\Adapter\Product\Update\ProductSupplierUpdater
         arguments:
             - '@prestashop.adapter.product.product_repository'
-            - '@prestashop.adapter.supplier.supplier_repository'
+            - '@prestashop.adapter.supplier.repository.supplier_repository'
             - '@prestashop.adapter.product.repository.product_supplier_repository'
