+++ src/PrestaShopBundle/Resources/config/services/adapter/supplier.yml
@@ -88,5 +88,5 @@ services:
       - name: 'tactician.handler'
         command: 'PrestaShop\PrestaShop\Core\Domain\Supplier\Command\EditSupplierCommand'
 
-  prestashop.adapter.supplier.supplier_repository:
-      class: 'PrestaShop\PrestaShop\Adapter\Supplier\SupplierRepository'
+  prestashop.adapter.supplier.repository.supplier_repository:
+      class: 'PrestaShop\PrestaShop\Adapter\Supplier\Repository\SupplierRepository'
